rune: Remove unwraps in languageserver
