+++ crates/rune/src/languageserver/envelope.rs
@@ -153,19 +153,3 @@ impl<'a> serde::Deserialize<'a> for V2 {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::Code;
-
-    #[test]
-    fn test_code() {
-        let code: Code = serde_json::from_str("-1").unwrap();
-        assert_eq!(code, Code::Unknown(-1));
-        assert_eq!(serde_json::to_string(&code).unwrap(), "-1");
-
-        let code: Code = serde_json::from_str("-32601").unwrap();
-        assert_eq!(code, Code::MethodNotFound);
-        assert_eq!(serde_json::to_string(&code).unwrap(), "-32601");
-    }
-}
